Правки решений

+++ src/main/java/org/study/javarush/java/core/level02/tasks02/f/Solution31.java
@@ -9,7 +9,7 @@ public class Solution31 {
         int firstPrice = scanner.nextInt();
         int secondPrice = scanner.nextInt();
 
-        int totalPrice = firstPrice + secondPrice;
+        long totalPrice = (long) firstPrice + secondPrice;
 
         System.out.println(totalPrice);
     }
