+++ src/main/java/org/study/javarush/java/core/level04/tasks04/e/Solution68.java
@@ -6,7 +6,7 @@ public class Solution68 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int totalSales = 0;
+        long totalSales = 0;
 
         while (true) {
             int sale = scanner.nextInt();
